[*] Move enumeration helper functions to device wrapper
