+++ include/inexor/vulkan-renderer/wrapper/device.hpp
@@ -14,6 +14,21 @@
 
 namespace inexor::vulkan_renderer::wrapper {
 
+/// Call vkEnumerateDeviceExtensionProperties
+/// @warning This can return an empty vector! It is the responsibility of the caller to check and account for this!
+/// @note Because device layers are deprecated in Vulkan, we are not exposing ``pLayerName`` parameter of
+/// ``vkEnumerateDeviceExtensionProperties`` as a parameter here
+/// @param layer_name The name of the (empty by default)
+/// @return A std::vector of all device extension properties of a physical device
+[[nodiscard]] std::vector<VkExtensionProperties>
+get_all_physical_device_extension_properties(VkPhysicalDevice physical_device);
+
+/// Call vkEnumeratePhysicalDevices
+/// @warning This can return an empty vector! It is the responsibility of the caller to check and account for this!
+/// @param inst The Vulkan instance
+/// @return A std::vector of all physical devices which are available on the system
+[[nodiscard]] std::vector<VkPhysicalDevice> get_all_physical_devices(VkInstance inst);
+
 /// Call vkGetPhysicalDeviceMemoryProperties
 /// @param physical_device The physical device
 /// @return The memory properties of the physical device
