+++ src/vulkan-renderer/wrapper/device.cpp
@@ -2,7 +2,6 @@
 
 #include "inexor/vulkan-renderer/exception.hpp"
 #include "inexor/vulkan-renderer/settings_decision_maker.hpp"
-#include "inexor/vulkan-renderer/vk_tools/enumerate.hpp"
 #include "inexor/vulkan-renderer/wrapper/make_info.hpp"
 
 #define VMA_IMPLEMENTATION
@@ -23,6 +22,7 @@
 
 #include <algorithm>
 #include <fstream>
+#include <utility>
 
 namespace {
 
@@ -32,6 +32,44 @@ constexpr float DEFAULT_QUEUE_PRIORITY = 1.0f;
 } // namespace
 
 namespace inexor::vulkan_renderer::wrapper {
+
+std::vector<VkExtensionProperties>
+get_all_physical_device_extension_properties(const VkPhysicalDevice physical_device) {
+    assert(physical_device);
+
+    // Query how many device extensions are available for the physical device
+    std::uint32_t count = 0;
+    // Because device layers are deprecated, we will never fill the pLayerName parameter
+    if (const auto result = vkEnumerateDeviceExtensionProperties(physical_device, nullptr, &count, nullptr);
+        result != VK_SUCCESS) {
+        throw VulkanException("Error: vkEnumerateDeviceExtensionProperties failed!", result);
+    }
+
+    /// Query device extension properties of the physical device
+    std::vector<VkExtensionProperties> extensions(count);
+    if (const auto result = vkEnumerateDeviceExtensionProperties(physical_device, nullptr, &count, extensions.data());
+        result != VK_SUCCESS) {
+        throw VulkanException("Error: vkEnumerateDeviceExtensionProperties failed!", result);
+    }
+    return std::move(extensions);
+}
+
+std::vector<VkPhysicalDevice> get_all_physical_devices(const VkInstance inst) {
+    assert(inst);
+
+    // Query how many physical devices are available on the system
+    std::uint32_t count = 0;
+    if (const auto result = vkEnumeratePhysicalDevices(inst, &count, nullptr); result != VK_SUCCESS) {
+        throw VulkanException("Error: vkEnumeratePhysicalDevices failed!", result);
+    }
+
+    // Query physical devices of the system
+    std::vector<VkPhysicalDevice> physical_devices(count);
+    if (const auto result = vkEnumeratePhysicalDevices(inst, &count, physical_devices.data()); result != VK_SUCCESS) {
+        throw VulkanException("Error: vkEnumeratePhysicalDevices failed!", result);
+    }
+    return std::move(physical_devices);
+}
 
 VkPhysicalDeviceMemoryProperties get_physical_device_memory_properties(const VkPhysicalDevice physical_device) {
     assert(physical_device);
@@ -58,7 +96,7 @@ bool is_extension_supported(const VkPhysicalDevice physical_device, const std::s
     assert(physical_device);
     assert(!extension_name.empty());
 
-    const auto extension_props = vk_tools::get_all_device_extension_properties(physical_device);
+    const auto extension_props = get_all_physical_device_extension_properties(physical_device);
     if (extension_props.size() == 0) {
         spdlog::info("No Vulkan device extensions available!");
         return false;
@@ -92,7 +130,7 @@ std::optional<VkPhysicalDevice> pick_graphics_card(const VkInstance inst, const 
     assert(inst);
     assert(surface);
 
-    auto physical_devices = vk_tools::get_all_physical_devices(inst);
+    auto physical_devices = get_all_physical_devices(inst);
     if (physical_devices.empty()) {
         throw std::runtime_error("Error: No physical devices available!");
     }
